+++ batchgenerators/dataloading/__init__.py
@@ -1,3 +1,4 @@
-from .data_loader import DataLoaderBase, SlimDataLoaderBase
+from .data_loader import DataLoaderBase, SlimDataLoaderBase, DataLoader, DataLoaderFromDataset
+from .dataset import Dataset
 from .multi_threaded_augmenter import MultiThreadedAugmenter
 from .single_threaded_augmenter import SingleThreadedAugmenter
